Unify Node and browser simulation orchestration
simulation-runner.ts and public/simulationRunner.ts were ~95% identical
copies (config validation, course resolution, race-condition parsing,
skill filtering, cost/stat aggregation) and had already drifted: the
browser hardcoded 500 simulations and dropped numSimulations from its
config type entirely. Extract all of it into a runtime-neutral
shared/simulation-orchestrator.ts; each runner now contributes only its
worker transport (worker_threads vs Web Worker) and concurrency source.

This fixes the browser ignoring numSimulations, makes the running-style
lookup fail loudly instead of silently defaulting to Pace Chaser, and
reports the actual per-skill sample count when per-combination batching
rounds it below the request.

Also fix parseAndBranch to intersect a field that appears twice in one
AND-branch (e.g. distance_type>=2&distance_type<=3) instead of letting the
second term overwrite the first, and export generateRepresentative.

Tests: shared orchestrator via a fake worker adapter (first coverage of
the browser code path), generateRepresentative apportionment, and the
parseAndBranch repeated-field cases.
